+++ tests/integration/standard/test_cluster.py
@@ -36,6 +36,8 @@ from cassandra import ConsistencyLevel
 
 from cassandra.pool import Host
 from cassandra.query import SimpleStatement, TraceUnavailable, tuple_factory
+from cassandra.auth import PlainTextAuthProvider, SaslAuthProvider
+from cassandra import connection
 
 from tests import notwindows
 from tests.integration import use_singledc, PROTOCOL_VERSION, get_server_versions, CASSANDRA_VERSION, \
@@ -671,6 +673,57 @@ class ClusterTests(unittest.TestCase):
         self.assertIn(query, str(future))
         self.assertIn('result', str(future))
         cluster.shutdown()
+
+    def test_can_connect_with_plainauth(self):
+        """
+        Verify that we can connect setting PlainTextAuthProvider against a
+        C* server without authentication set. We also verify a warning is
+        issued per connection. This test is here instead of in test_authentication.py
+        because the C* server running in that module has auth set.
+
+        @since 3.14
+        @jira_ticket PYTHON-940
+        @expected_result we can connect, query C* and warning are issued
+
+        @test_category auth
+        """
+        auth_provider = PlainTextAuthProvider(
+            username="made_up_username",
+            password="made_up_password"
+        )
+        self._warning_are_issued_when_auth(auth_provider)
+
+    def test_can_connect_with_sslauth(self):
+        """
+        Verify that we can connect setting SaslAuthProvider against a
+        C* server without authentication set. We also verify a warning is
+        issued per connection. This test is here instead of in test_authentication.py
+        because the C* server running in that module has auth set.
+
+        @since 3.14
+        @jira_ticket PYTHON-940
+        @expected_result we can connect, query C* and warning are issued
+
+        @test_category auth
+        """
+        sasl_kwargs = {'service': 'cassandra',
+                       'mechanism': 'PLAIN',
+                       'qops': ['auth'],
+                       'username': "made_up_username",
+                       'password': "made_up_password"}
+
+        auth_provider = SaslAuthProvider(**sasl_kwargs)
+        self._warning_are_issued_when_auth(auth_provider)
+
+    def _warning_are_issued_when_auth(self, auth_provider):
+        with MockLoggingHandler().set_module_name(connection.__name__) as mock_handler:
+            with Cluster(auth_provider=auth_provider) as cluster:
+                session = cluster.connect()
+                self.assertIsNotNone(session.execute("SELECT * from system.local"))
+
+            # Three conenctions to nodes plus the control connection
+            self.assertEqual(4, mock_handler.get_message_count('warning',
+                                                               "An authentication challenge was not sent"))
 
     def test_idle_heartbeat(self):
         interval = 2